sink/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
attrs==24.2.0
blinker==1.8.2
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
colorzero==2.0
cryptography==44.0.0
frozenlist==1.4.1
gpiozero==2.0.1
idna==3.7
iso8601==2.1.0
multidict==6.0.5
ndg-httpsclient==0.5.1
paho-mqtt==2.1.0
psutil==6.0.0
pusher==3.3.2
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyNaCl==1.5.0
pyOpenSSL==24.3.0
pypika-tortoise==0.2.1
Pysher==1.0.8
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
RPi.GPIO==0.7.1
setproctitle==1.3.3
setuptools==75.3.0
six==1.17.0
tortoise-orm==0.21.7
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
yarl==1.9.4
